장바구니

+++ routes/goods.js
@@ -1,6 +1,27 @@
 const express = require('express');
 const router = express.Router();
 const Goods = require('../schemas/goods');
+const Cart = require('../schemas/cart');
+const authMiddleware = require('../middlewares/auth-middleware');
+
+router.get('/goods/cart', authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
+  const carts = await Cart.find({ userId });
+  // [{goodsId, quantity}]
+  const goodsIds = carts.map((cart) => {
+    return cart.goodsId;
+  });
+  const goods = await Goods.find({ goodsId: goodsIds });
+  const results = carts.map((cart) => {
+    return {
+      quantity: cart.quantity,
+      goods: goods.find((item) => item.goodsId === cart.goodsId),
+    };
+  });
+  res.json({
+    carts: results,
+  });
+});
 
 //GET, 모든 상품목록 goods 정보 가져오기
 router.get('/goods', async (req, res) => {
@@ -77,12 +98,12 @@ router.get('/goods/:goodsId', async (req, res) => {
 //     },
 //   ];
 
-const Cart = require('../schemas/cart');
-router.post('/goods/:goodsId/cart', async (req, res) => {
+router.post('/goods/:goodsId/cart', authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     return res.status(400).json({
       success: false,
@@ -90,31 +111,33 @@ router.post('/goods/:goodsId/cart', async (req, res) => {
     });
   }
 
-  await Cart.create({ goodsId, quantity });
+  await Cart.create({ userId, goodsId, quantity });
 
   res.json({ result: 'success' });
 });
 
-router.put('/goods/:goodsId/cart', async (req, res) => {
+router.put('/goods/:goodsId/cart', authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
   const { quantity } = req.body;
 
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
   if (existsCarts.length) {
     await Cart.updateOne(
-      { goodsId: goodsId },
+      { userId, goodsId: goodsId },
       { $set: { quantity: quantity } }
     );
   }
   res.status(200).json({ success: true });
 });
 
-router.delete('/goods/:goodsId/cart', async (req, res) => {
+router.delete('/goods/:goodsId/cart', authMiddleware, async (req, res) => {
+  const { userId } = res.locals.user;
   const { goodsId } = req.params;
-  const existsCarts = await Cart.find({ goodsId });
+  const existsCarts = await Cart.find({ userId, goodsId });
 
   if (existsCarts.length) {
-    await Cart.deleteOne({ goodsId: goodsId });
+    await Cart.deleteOne({ userId, goodsId: goodsId });
   }
 
   res.json({ result: 'success' });
